move retry file around
