--- python/ray/util/tests/unit/BUILD
@@ -1,10 +1,0 @@
-load("//bazel:python.bzl", "py_test_module_list")
-
-py_test_module_list(
-  files = [
-    "test_retry.py",
-  ],
-  size = "small",
-  tags = ["team:core"],
-  deps = [],
-)
